Rename to depositmonitor2
