+++ src/deposit-monitor-manager2.d.ts
