+++ src/index.js
@@ -3,7 +3,7 @@ function __export(m) {
     for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
 }
 Object.defineProperty(exports, "__esModule", { value: true });
-__export(require("./deposit-monitor"));
-__export(require("./deposit-monitor-manager"));
+__export(require("./deposit-monitor2"));
+__export(require("./deposit-monitor-manager2"));
 __export(require("./schema"));
 //# sourceMappingURL=index.js.map
