+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './deposit-monitor'
-export * from './deposit-monitor-manager'
+export * from './deposit-monitor2'
+export * from './deposit-monitor-manager2'
 export * from './types'
 export * from './schema'
